ebpf structure refactor
