+++ huginn-proxy/src/lib.rs
@@ -1,5 +1,3 @@
 #![forbid(unsafe_code)]
 
-#[cfg(feature = "ebpf-tcp")]
-#[doc(hidden)]
-pub mod ebpf_config;
+pub mod ebpf;
