+++ huginn-proxy/tests/ebpf/main.rs
@@ -1,0 +1,3 @@
+#![cfg(feature = "ebpf-tcp")]
+
+mod config;
